requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
python-dotenv==0.19.0
psycopg2-binary==2.9.1
python-jwt==3.3.0
werkzeug==2.0.1